Add files via upload
Self-hypnosis is a technique where you put yourself in a deeply relaxed state. It balances the conscious and sub-conscious minds, where the sub-conscious mind is awakened and the conscious mind is put into a quiet state. But, have you wondered why? Our conscious mind is responsible for logical thinking and decision making whereas, the subconscious mind is responsible for storing memories, beliefs and habits that is why by quieting the conscious mind you can access the subconscious mind more effectively.